+++ src/weight_measurement.c
@@ -11,14 +11,14 @@
 #include <STM32_bsp/constants.h>
 #include <STM32_bsp/gpio.h>
 
-float r_wheatstone_bridge()
+uint16_t r_wheatstone_bridge()
 {
   float output;
-  float mcu_vdd = calculate_MCU_Vdd();
+  uint16_t mcu_vdd = calculate_MCU_Vdd();
   uint16_t digital_val;
 
   digital_val = r_single_ext_channel_ADC(STRAIN_GAUGE);
-  output = (mcu_vdd/4095) * digital_val;
+  output = (mcu_vdd/4095.0) * digital_val;
 
   return output;
 }
